requirements.txt
Flask>=2.3.0
Flask-SQLAlchemy>=3.0.0
Flask-WTF>=1.1.0
Flask-Login>=0.6.0
Flask-Migrate>=4.0.0
Flask-Mail>=0.9.1
WTForms>=3.0.0
Werkzeug>=2.3.0
python-dotenv>=1.0.0
email-validator>=2.0.0
bcrypt>=4.0.0
python-slugify>=8.0.0
Markdown>=3.4.0
requests>=2.31.0
aiohttp>=3.8.0